Auto Updating Command Data Feature Implemented
Some commands have been added. But the biggest addition for this commit is the command data being "off-site". All the data is stored in MongoDB, which allows us to update the command data without having to take down the bot.

I have an idea where I want to be able to store the invoke functions of the commands in the mongodb as well, which allows us to modify and edit the command without having to have the bot to be down. Its a thought, but will need research in order for it to work.