Fully dynamic help command
Completely made the help command to be entirely dynamic over static which it used to be. Also optimized the command data in the database to work flawlessly with the command invokes.

Next step is working on the rest of the fun commands, and working on any systems like currency or levels.