Radio Command works, but its still needs work
Just finished making the radio command. All the stations are in the database, so we can add or remove any stations without turning off the bot. There are still some bugs and messages that need to be worked out.

Also moved all the animal related commands into one command called animal, except for fish, as that is something different.